Retry failed external log entries
If the system hsa no network, or the network connection fails, the
external log will be attempted again after a delay.

The delay increases exponentially up to 5 hours.
The maximum number of stored entries to queue up is 100.

The failed entries will not survive a system crash or reboot as they
are not persisted to disk.